self clearing fixtures for repo and database
The scope of the test_app fixture extended
for the whole session.

This means that the status of the fixture
is not reset between tests and, in fact,
some test in test_edits relied one on each
other state modifications: if they were
run separately they failed.

The fixture has been split in topic levels
(git, database, erp) and it has been. And
a try/finally schema has been used to ensure
clearing them. And scope has been changed
to function.

TODO: explore if we could keep the database
creation at session scope by providing a
transaction fixture at function level,
like we did with ERP.