single parameter set for save_template
Do not force save_template to accept two
different set of parameters. Just do the
parameter mapping in upload_edit.

This simplifies testing and logic to come
for translation management.

Prefered def_text_body over text_body
because of simmetry with load_template
and because is a direct mapping with
translation parameters.